pyproject.toml
[project]
name = "vulnerable-webapp"
version = "0.1.0"
description = "Vulnerable web application for AWS security testing"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "boto3>=1.34.0",
    "pymysql>=1.1.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "pyjwt>=2.8.0",
    "sqlalchemy>=2.0.23",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
